# docs/swagger.yaml
definitions:
...
  DTO.Client:
    properties:
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      name:
        example: John
        type: string
      phone:
        example: "+15555555555"
        type: string
      surname:
        example: Doe
        type: string
      verified:
        example: false
        type: boolean
    type: object
...
  DTO.CreateClient:
    properties:
      email:
        example: john.doe@example.com
        type: string
      name:
        example: John
        type: string
      password:
        example: 1SecurePswd!
        type: string
      phone:
        example: "+15555555555"
        type: string
      surname:
        example: Doe
        type: string
    type: object
...
  DTO.ErrorResponse:
    description: Error Response Model
    properties:
      message:
        example: Not Found
        type: string
    type: object
...
  DTO.LoginByEmailCode:
    properties:
      code:
        example: "123456"
        type: string
      email:
        example: john.doe@example.com
        type: string
    type: object
  DTO.LoginClient:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: 1SecurePswd!
        type: string
    type: object
...
  DTO.PasswordReseted:
    properties:
      password:
        type: string
    type: object
...
  dJSON.ClientAppointment:
    properties:
      branch_address:
        example: 76, Example street, My city, My country, 09090790
        type: string
      company_id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      company_name:
        example: Company name example
        type: string
      service_name:
        example: Service name example
        type: string
      start_time:
        example: "2021-01-01T09:00:00Z"
        type: string
    type: object
...
  dJSON.Image:
    properties:
      alt:
        example: Image of something
        type: string
      caption:
        example: This is the image we talk about
        type: string
      title:
        example: Title of this image
        type: string
      url:
        example: https://example.com/image.png
        type: string
    type: object
  dJSON.Images:
    properties:
      background:
        $ref: '#/definitions/dJSON.Image'
      banner:
        $ref: '#/definitions/dJSON.Image'
      favicon:
        $ref: '#/definitions/dJSON.Image'
      logo:
        $ref: '#/definitions/dJSON.Image'
      profile:
        $ref: '#/definitions/dJSON.Image'
    type: object
...
info:
  contact: {}
  description: Swagger API for testing and debugging
  termsOfService: http://swagger.io/terms/
  title: GO-Fiber API
  version: "1.0"
paths:
...
  /client:
    post:
      consumes:
      - application/json
      description: Create an client
      parameters:
      - description: Client
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/DTO.CreateClient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      summary: Create client
      tags:
      - Client
  /client/{id}:
    delete:
      description: Delete an client
      parameters:
      - description: X-Auth-Token
        in: header
        name: X-Auth-Token
        required: true
        type: string
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Delete client
      tags:
      - Client
    get:
      description: Retrieve an client by its ID
      parameters:
      - description: X-Auth-Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get client by ID
      tags:
      - Client
    patch:
      consumes:
      - application/json
      description: Update an client
      parameters:
      - description: X-Auth-Token
        in: header
        name: X-Auth-Token
        required: true
        type: string
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Client
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/DTO.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Update client
      tags:
      - Client
  /client/{id}/appointments:
    get:
      description: Get only the appointments field from a client
      parameters:
      - description: X-Auth-Token
        in: header
        name: X-Auth-Token
        required: true
        type: string
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dJSON.ClientAppointment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get client appointments
      tags:
      - Client
  /client/{id}/design/images:
    patch:
      consumes:
      - application/json
      description: Update the design images of an client
      parameters:
      - description: X-Auth-Token
        in: header
        name: X-Auth-Token
        required: true
        type: string
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Profile image
        in: formData
        name: profile
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Update client design images
      tags:
      - Client
  /client/{id}/design/images/{image_type}:
    delete:
      description: Delete the design images of an client
      parameters:
      - description: X-Auth-Token
        in: header
        name: X-Auth-Token
        required: true
        type: string
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Image Type
        in: path
        name: image_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dJSON.Images'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Delete client design images
      tags:
      - Client
  /client/email/{email}:
    get:
      description: Retrieve an client by its email
      parameters:
      - description: Client Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      summary: Get client by email
      tags:
      - Client
  /client/login:
    post:
      consumes:
      - application/json
      description: Log in an client using password
      parameters:
      - description: Client
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/DTO.LoginClient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      summary: Login
      tags:
      - Client
  /client/login-with-code:
    post:
      consumes:
      - application/json
      description: Login client using email and validation code
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/DTO.LoginByEmailCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      summary: Login client by email code
      tags:
      - client/auth
  /client/reset-password/{email}:
    post:
      description: Reset the password of a client by its email
      parameters:
      - description: X-Auth-Token
        in: header
        name: X-Auth-Token
        required: true
        type: string
      - description: Client Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.PasswordReseted'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Reset client password by email
      tags:
      - Client
  /client/send-login-code/email/{email}:
    post:
      description: Send a login validation code to a client's email
      parameters:
      - description: Client Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      summary: Send client login validation code by email
      tags:
      - Client
...
securityDefinitions:
  ApiKeyAuth:
    description: |-
      Enter the token in the format: <token>
      Enter the company ID in the format: <company_id>
    in: header
    name: X-Company-ID
    type: apiKey
swagger: "2.0"